// packages/contracts/src/proto/crypto/v1/envelope_pb.ts
// @generated by protoc-gen-es v2.14.0 with parameter "target=ts,import_extension=.js"
// @generated from file crypto/v1/envelope.proto (package crypto.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file crypto/v1/envelope.proto.
 */
export const file_crypto_v1_envelope: GenFile = /*@__PURE__*/
  fileDesc("ChhjcnlwdG8vdjEvZW52ZWxvcGUucHJvdG8SCWNyeXB0by52MSJpCg5TaWduZWRFbnZlbG9wZRIRCglzZW5zb3JfaWQYASABKAwSEQoJdGltZXN0YW1wGAIgASgEEg0KBW5vbmNlGAMgASgMEg8KB21lc3NhZ2UYBCABKAwSEQoJc2lnbmF0dXJlGAUgASgMIj8KE1NpZ25lZEVudmVsb3BlQmF0Y2gSKAoFYmF0Y2gYASADKAsyGS5jcnlwdG8udjEuU2lnbmVkRW52ZWxvcGViBnByb3RvMw");

/**
 * @generated from message crypto.v1.SignedEnvelope
 */
export type SignedEnvelope = Message<"crypto.v1.SignedEnvelope"> & {
  /**
   * @generated from field: bytes sensor_id = 1;
   */
  sensorId: Uint8Array;

  /**
   * @generated from field: uint64 timestamp = 2;
   */
  timestamp: bigint;

  /**
   * @generated from field: bytes nonce = 3;
   */
  nonce: Uint8Array;

  /**
   * @generated from field: bytes message = 4;
   */
  message: Uint8Array;

  /**
   * @generated from field: bytes signature = 5;
   */
  signature: Uint8Array;
};

/**
 * Describes the message crypto.v1.SignedEnvelope.
 * Use `create(SignedEnvelopeSchema)` to create a new message.
 */
export const SignedEnvelopeSchema: GenMessage<SignedEnvelope> = /*@__PURE__*/
  messageDesc(file_crypto_v1_envelope, 0);

/**
 * @generated from message crypto.v1.SignedEnvelopeBatch
 */
export type SignedEnvelopeBatch = Message<"crypto.v1.SignedEnvelopeBatch"> & {
  /**
   * @generated from field: repeated crypto.v1.SignedEnvelope batch = 1;
   */
  batch: SignedEnvelope[];
};

/**
 * Describes the message crypto.v1.SignedEnvelopeBatch.
 * Use `create(SignedEnvelopeBatchSchema)` to create a new message.
 */
export const SignedEnvelopeBatchSchema: GenMessage<SignedEnvelopeBatch> = /*@__PURE__*/
  messageDesc(file_crypto_v1_envelope, 1);